@@ -8,7 +8,6 @@
     "build": "gulp default",
     "production": "NODE_ENV=production gulp default",
     "clean": "del dist",
-    "precommit": "lint-staged",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "lint-staged": {
@@ -23,48 +22,47 @@
   "author": "uhyo",
   "license": "MIT",
   "devDependencies": {
-    "@types/react": "^16.4.11",
-    "@types/react-dom": "^16.0.7",
-    "autoprefixer": "^9.1.3",
+    "@types/react": "^16.8.6",
+    "@types/react-dom": "^16.8.2",
+    "autoprefixer": "^9.4.9",
     "awesome-typescript-loader": "^5.2.0",
     "connect": "^3.6.6",
     "connect-livereload": "^0.6.0",
-    "css-loader": "^1.0.0",
+    "css-loader": "^2.1.0",
     "del": "^3.0.0",
-    "file-loader": "^2.0.0",
+    "file-loader": "^3.0.1",
     "gulp": "^4.0.0",
     "gulp-changed": "^3.2.0",
     "gulp-if": "^2.0.2",
     "gulp-livereload": "^4.0.1",
     "gulp-replace": "^1.0.0",
-    "gulp-sourcemaps": "^2.6.4",
-    "gulp-typed-css-modules": "^1.1.0",
-    "gulp-uglify": "^3.0.1",
-    "husky": "^0.14.3",
+    "gulp-sourcemaps": "^2.6.5",
+    "gulp-typed-css-modules": "^2.0.1",
+    "gulp-uglify": "^3.0.2",
+    "husky": "^1.3.1",
     "img-loader": "^3.0.0",
-    "lint-staged": "^7.2.2",
-    "masao": "^0.1.12",
+    "lint-staged": "^8.1.5",
+    "masao": "^0.2.0",
     "npm-run-all": "^4.1.3",
     "postcss-loader": "^3.0.0",
-    "postcss-nested": "^3.0.0",
+    "postcss-nested": "^4.1.2",
     "prettier": "^1.14.2",
-    "react-redux": "^5.0.7",
+    "react-redux": "^6.0.1",
     "redux": "^4.0.0",
-    "redux-logic": "^0.15.4",
-    "rxjs": "^5.5.11",
+    "redux-logic": "^2.1.1",
     "serve-static": "^1.13.2",
     "source-map-loader": "^0.2.4",
-    "style-loader": "^0.22.1",
-    "typescript": "^3.0.1",
+    "style-loader": "^0.23.1",
+    "typescript": "^3.3.3333",
     "uglify-es": "^3.3.9",
-    "uglifyjs-webpack-plugin": "^1.3.0",
+    "uglifyjs-webpack-plugin": "^2.1.2",
     "url-loader": "^1.1.0",
-    "webpack": "^4.16.5"
+    "webpack": "^4.29.6"
   },
   "dependencies": {
-    "@types/react-redux": "^6.0.6",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2"
+    "@types/react-redux": "^7.0.1",
+    "react": "^16.8.3",
+    "react-dom": "^16.8.3"
   },
   "repository": {
     "type": "git",
@@ -73,5 +71,10 @@
   "bugs": {
     "url": "https://github.com/uhyo/masao-editor-standalone/issues"
   },
-  "homepage": "https://github.com/uhyo/masao-editor-standalone#readme"
+  "homepage": "https://github.com/uhyo/masao-editor-standalone#readme",
+  "husky": {
+    "hooks": {
+      "pre-commit": "lint-staged"
+    }
+  }
 }
